add "auto-select pictures" for albums
The auto-selecting is done automatically for albums of newly added songs
It can also be selected from album context menu when multiple albums are selected

The way the covers are auto-selected is not currently configurable.
It first checks for embedded covers (a bit slow if files not cached).
Then select one picture in one of the folders of the album or in the common parent folder. The folder must contain only songs from the album.
Picture files containing words from the album name or from cover,front,folder,thumb,thumbnail are more likely to be picked