mpv: fix gapless playback with versions >=0.28.0
`tracks-changed` event fires more frequently since 0.28, breaking the
logic here so that all track changes are treated as user-initiated.
The event has also been deprecated for some time, so its behavior is
officially undefined, and we're not supposed to use it.

Instead, we now use the `file-loaded` event which always neatly fires
once per track transition.

Changes should be backwards-compatible (tested with mpv 0.27)